Refactor waiver controller to waiver_admin
* Opens up edit/update actions for owners in waiver controller
* Owners need to be able to update the player to drop
* Moved waiver_admin route behind authorize_admin plug
** Eliminates conflict with multiple edit routes on Waiver resource
** Wasn't sure how to differentiate in the abilities.ex file
* Enables issue #113 resolution
