+++ web/controllers/waiver_controller.ex
@@ -10,10 +10,6 @@ defmodule Ex338.WaiverController do
     preload: [:owners], persisted: true, id_name: "fantasy_team_id",
     unauthorized_handler: {Authorization, :handle_unauthorized}
 
-  plug :load_and_authorize_resource, model: Waiver, only: [:edit, :update],
-    preload: [:fantasy_team, :add_fantasy_player, :drop_fantasy_player],
-    unauthorized_handler: {Authorization, :handle_unauthorized}
-
   def index(conn, %{"fantasy_league_id" => league_id}) do
     fantasy_league = FantasyLeague |> Repo.get(league_id)
 
@@ -82,29 +78,6 @@ defmodule Ex338.WaiverController do
                                  fantasy_league: fantasy_league,
                                  owned_players: owned_players,
                                  avail_players: avail_players)
-    end
-  end
-
-  def edit(conn, _) do
-    waiver    = conn.assigns.waiver
-    changeset = Waiver.changeset(waiver)
-
-    render(conn, "edit.html", waiver: waiver, changeset: changeset)
-  end
-
-  def update(conn,%{"id" => _, "waiver" => params}) do
-    waiver = conn.assigns.waiver
-
-    result = Waiver.update_waiver(waiver, params)
-
-    case result do
-      {:ok,  %{waiver: _waiver}} ->
-        conn
-        |> put_flash(:info, "Waiver successfully processed")
-        |> redirect(to: fantasy_league_waiver_path(conn, :index,
-                        waiver.fantasy_team.fantasy_league_id))
-      {:error, _, changeset, _} ->
-        render(conn, "edit.html", waiver: waiver, changeset: changeset)
     end
   end
 end
