+++ web/views/waiver_admin_view.ex
@@ -1,0 +1,3 @@
+defmodule Ex338.WaiverAdminView do
+  use Ex338.Web, :view
+end
